move last function from helper_functions_legacy
