+++ resurfemg/helper_functions_legacy.py
@@ -562,51 +562,51 @@ from .preprocessing.envelope import naive_rolling_rms
 #     return ecg_component
 
 
-def working_pipeline_exp(our_chosen_file):
-    """This function is legacy.
-    It produces a filtered respiratory EMG signal from a
-    3 lead sEMG file. A better
-    option is a corresponding function in multi_lead_type
-    The inputs are :code:`our_chosen_file` which we
-    give the function as a string of filename.  The output is the
-    processed EMG signal filtered and seperated from ECG components.
-    The algorithm to pick out the EMG here is by having
-    more peaks.
+# def working_pipeline_exp(our_chosen_file):
+#     """This function is legacy.
+#     It produces a filtered respiratory EMG signal from a
+#     3 lead sEMG file. A better
+#     option is a corresponding function in multi_lead_type
+#     The inputs are :code:`our_chosen_file` which we
+#     give the function as a string of filename.  The output is the
+#     processed EMG signal filtered and seperated from ECG components.
+#     The algorithm to pick out the EMG here is by having
+#     more peaks.
 
-    :param our_chosen_file: Poly5 file
-    :type our_chosen_file: ~TMSiSDK.file_readers.Poly5Reader
+#     :param our_chosen_file: Poly5 file
+#     :type our_chosen_file: ~TMSiSDK.file_readers.Poly5Reader
 
-    :returns: final_envelope_a
-    :rtype: ~numpy.ndarray
-    """
-    cut_file_data = bad_end_cutter(
-        our_chosen_file,
-        percent_to_cut=3,
-        tolerance_percent=5,
-    )
-    bd_filtered_file_data = emg_bandpass_butter_sample(
-        cut_file_data,
-        5,
-        450,
-        2048,
-        output='sos',
-    )
-    # end-cutting again to get rid of filtering artifacts
-    re_cut_file_data = bad_end_cutter_for_samples(
-        bd_filtered_file_data,
-        percent_to_cut=3,
-        tolerance_percent=5,
-    )
-    # do ICA
-    components = compute_ICA_two_comp(re_cut_file_data)
-    #  pick components with more peak
-    emg = pick_more_peaks_array(components)
-    # now process it in final steps
-    abs_values = abs(emg)
-    final_envelope_d = emg_highpass_butter(abs_values, 150, 2048)
-    final_envelope_a = naive_rolling_rms(final_envelope_d, 300)
+#     :returns: final_envelope_a
+#     :rtype: ~numpy.ndarray
+#     """
+#     cut_file_data = bad_end_cutter(
+#         our_chosen_file,
+#         percent_to_cut=3,
+#         tolerance_percent=5,
+#     )
+#     bd_filtered_file_data = emg_bandpass_butter_sample(
+#         cut_file_data,
+#         5,
+#         450,
+#         2048,
+#         output='sos',
+#     )
+#     # end-cutting again to get rid of filtering artifacts
+#     re_cut_file_data = bad_end_cutter_for_samples(
+#         bd_filtered_file_data,
+#         percent_to_cut=3,
+#         tolerance_percent=5,
+#     )
+#     # do ICA
+#     components = compute_ICA_two_comp(re_cut_file_data)
+#     #  pick components with more peak
+#     emg = pick_more_peaks_array(components)
+#     # now process it in final steps
+#     abs_values = abs(emg)
+#     final_envelope_d = emg_highpass_butter(abs_values, 150, 2048)
+#     final_envelope_a = naive_rolling_rms(final_envelope_d, 300)
 
-    return final_envelope_a
+#     return final_envelope_a
 
 
 # def working_pipeline_pre_ml(our_chosen_samples, picker='heart'):
